Pacify a GitHub CodeQL warning.

+++ bzlib.c
@@ -103,7 +103,7 @@ int bz_config_ok ( void )
 static
 void* default_bzalloc ( void* opaque, Int32 items, Int32 size )
 {
-   void* v = malloc ( items * size );
+   void* v = malloc ( items * (size_t)size );
    (void) opaque;
    return v;
 }
